% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_sf_2020}
\alias{block_sf_2020}
\title{Census Blocks (2020 vintage)}
\format{
A multipolygon \code{sf} object with 144082 rows and
4 variables:
\describe{
\item{geoid_block}{Unique 15-digit block ID, assigned by the Census Bureau.
The parent tract and block group can be identified from the first 11 and 12
digits, respectively. Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the block is in.
Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
block_sf_2020
}
\description{
The Census Blocks within the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2020 vintage. \strong{Use this version for data from the 2020 decennial census. For
data from the 2010 decennial census or the American Community Survey (ACS)
from 2010 through 2019, use \code{block_sf} (which will be replaced by this
dataset once the 2016-2020 ACS 5-year data is published).}
}
\details{
Census Bureau description:

\emph{"Blocks are statistical areas bounded by visible features, such as streets,
roads, streams, and railroad tracks, and by nonvisible boundaries, such as
selected property lines and city, township, school district, and county
limits and short line-of-sight extensions of streets and roads. Generally,
census blocks are small in area; for example, a block in a city bounded on
all sides by streets. Census blocks in suburban and rural areas may be large,
irregular, and bounded by a variety of features, such as roads, streams, and
transmission lines. In remote areas, census blocks may encompass hundreds of
square miles. Census blocks cover the entire territory of the United States,
Puerto Rico, and the Island Areas. Census blocks nest within all other
tabulated census geographic entities and are the basis for all tabulated
data."}
}
\examples{
# Display the blocks with ggplot2
library(ggplot2)
ggplot(block_sf_2020) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
