% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmap_mpa_sf}
\alias{cmap_mpa_sf}
\title{CMAP Metropolitan Planning Area}
\format{
A polygon \code{sf} object with 9 rows and
5 variables:
\describe{
\item{label_name}{County name (or township names, for partial counties).
Character.}
\item{county_fips}{Unique 5-digit county FIPS code. Character.}
\item{whole_county}{Does feature represent an entire county? Logical.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
\href{https://datahub.cmap.illinois.gov/dataset/cmap-metropolitan-planning-area-september-2014}{CMAP Data Hub}
}
\usage{
cmap_mpa_sf
}
\description{
This dataset depicts the individual counties (some only partial) that compose
the Chicago Metropolitan Planning Area, approved by the Board of
\href{https://www.cmap.illinois.gov}{CMAP} and the
\href{https://www.cmap.illinois.gov/committees/policy/mpo}{MPO Policy Committee}
in March of 2013. Final approval by Governor Quinn was granted on 11
September, 2014. The revised MPA includes the urbanized area established by
the US Census Bureau pursuant to the 2010 Decennial Census. The MPA was
expanded to include Somonauk and Sandwich Townships in DeKalb County.
}
\examples{
# Display the CMAP MPA counties with ggplot2
library(ggplot2)
ggplot(data = cmap_mpa_sf) + geom_sf(aes(fill = whole_county), lwd = 0.1) + theme_void()
}
\keyword{datasets}
