% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapgeo.R
\docType{package}
\name{cmapgeo}
\alias{cmapgeo}
\alias{cmapgeo-package}
\title{cmapgeo}
\description{
This package is a repository of geodata files representing various geographic
and administrative boundaries in the Chicago region, maintained by the
Chicago Metropolitan Agency for Planning (CMAP). Data is in the \code{sf} (simple
features) format.
}
\details{
Detailed documentation can be viewed at
\url{https://cmap-repos.github.io/cmapgeo}.

Please report issues and suggest improvements at
\url{https://github.com/CMAP-REPOS/cmapgeo/issues}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmap-repos.github.io/cmapgeo}
  \item \url{https://github.com/CMAP-REPOS/cmapgeo}
  \item Report bugs at \url{https://github.com/CMAP-REPOS/cmapgeo/issues}
}

}
\author{
\strong{Maintainer}: Noel Peterson \email{npeterson@cmap.illinois.gov} (\href{https://orcid.org/0000-0003-1290-2362}{ORCID})

Other contributors:
\itemize{
  \item Chicago Metropolitan Agency for Planning [copyright holder]
}

}
\keyword{internal}
