% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congress_sf}
\alias{congress_sf}
\title{U.S. Congressional Districts}
\format{
A multipolygon \code{sf} object with 17 rows and
6 variables:
\describe{
\item{dist_num}{Congressional District number. Integer.}
\item{dist_name}{Name of the district (full). Character.}
\item{dist_name_short}{Name of the district (short). Character.}
\item{cmap}{Does the district overlap the 7-county CMAP region? Logical.}
\item{sqmi}{Area in square miles. Double.} \item{geometry}{Feature
geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
congress_sf
}
\description{
The United States Congressional Districts in the state of Illinois. From the
US Census Bureau's TIGER/Line shapefiles, 2023 vintage. \strong{These districts
represent new district boundaries for the 2022 election (for the 118th
Congress).}
}
\details{
Census Bureau description:

\emph{"Congressional Districts are the 435 areas from which people are elected to
the U.S. House of Representatives. After the apportionment of congressional
seats among the states based on decennial census population counts, each
state with multiple seats is responsible for establishing congressional
districts for the purpose of electing representatives. Each congressional
district is to be as equal in population to all other congressional districts
in a state as practicable. For the District of Columbia, Puerto Rico, and
each Island Area, a separate code is used to identify the entire areas of
these state-equivalent entities as having a single nonvoting delegate."}
}
\examples{
# Display the Congressional Districts with ggplot2
library(ggplot2)
ggplot(congress_sf) + geom_sf(aes(fill = cmap), lwd = 0.1) + theme_void()
}
\keyword{datasets}
