% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapgeo.R
\docType{data}
\name{county_fips}
\alias{county_fips}
\title{County FIPS Code Lists}
\format{
An object of class \code{list} of length 4.
}
\usage{
county_fips
}
\description{
A named list comprising four named vectors, each of which contains the FIPS
codes (as 5-character strings) for the counties in one of the four primary
geographic divisions of CMAP's travel model and socioeconomic forecasts: the
7-county CMAP region, as well as the external counties in Illinois, Indiana,
and Wisconsin.
}
\examples{
county_fips$cmap  # 7-county CMAP region
county_fips$xil   # External Illinois counties
county_fips$xin   # External Indiana counties
county_fips$xwi   # External Wisconsin counties

# Get the names of the CMAP counties
names(county_fips$cmap)

# Combine all 4 divisions into a single vector
unlist(county_fips)

}
\keyword{datasets}
