% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapgeo.R
\docType{data}
\name{county_fips_codes}
\alias{county_fips_codes}
\title{County FIPS Code Lists}
\format{
An object of class \code{list} of length 5.
}
\usage{
county_fips_codes
}
\description{
A named list comprising four named vectors, each of which contains the FIPS
codes (as 5-character strings) for the counties in one of the four primary
geographic divisions of CMAP's travel model and socioeconomic forecasts: the
7-county CMAP region, as well as the external counties in Illinois, Indiana,
and Wisconsin.
}
\examples{
county_fips_codes$cmap  # 7-county CMAP region
county_fips_codes$xil   # Travel model external Illinois counties
county_fips_codes$xin   # Travel model external Indiana counties
county_fips_codes$xwi   # Travel model external Wisconsin counties
county_fips_codes$msa   # Chicago-Naperville-Elgin, IL-IN-WI MSA counties

# Get the names of the CMAP counties
names(county_fips_codes$cmap)

# Combine the 4 travel model divisions into a single vector
unlist(county_fips_codes[c("cmap", "xil", "xin", "xwi")])

}
\keyword{datasets}
