% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_sf}
\alias{county_sf}
\title{Counties}
\format{
A polygon \code{sf} object with 23 rows and
8 variables:
\describe{
\item{geoid_county}{Unique 5-digit county ID (a.k.a. FIPS code), assigned
by the Census Bureau. Character.}
\item{county}{County name. Character.}
\item{state}{2-letter state abbreviation. Character.}
\item{cmap}{Is the county part of the 7-county CMAP region? Logical.}
\item{msa}{Is the county part of the "Chicago-Naperville-Elgin, IL-IN-WI"
Metropolitan Statistical Area? Logical.}
\item{travel_model}{Is the county at least partially covered by CMAP's
travel modeling area? Logical.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
county_sf
}
\description{
The counties that are within the CMAP travel modeling area \strong{or} the
"Chicago-Naperville-Elgin, IL-IN-WI" Metropolitan Statistical Area (as
defined by the United States Office of Management and Budget). From the US
Census Bureau's TIGER/Line shapefiles, 2021 vintage.
}
\details{
Census Bureau description:

\emph{"The primary legal divisions of most states are termed counties. Each county
or statistically equivalent entity is assigned a three-character numeric
Federal Information Processing Series (FIPS) code based on alphabetical
sequence that is unique within state, and an eight-digit National Standard
(NS) code."}

Note: The Illinois counties of LaSalle, Lee and Ogle are included in their
entirety, although only portions of these counties are part of the CMAP
travel modeling area. The precise geographic extent of the CMAP travel
modeling area is reflected in \code{zone_sf} and \code{subzone_sf}.
}
\examples{
# Display the counties with ggplot2
library(ggplot2)
ggplot(county_sf) +
  geom_sf(aes(fill=paste(travel_model, msa)), lwd = 0.1) +
  geom_sf(data = dplyr::filter(county_sf, cmap), fill = NA, lwd = 0.5) +
  theme_void()
}
\keyword{datasets}
