% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eda_sf}
\alias{eda_sf}
\title{CMAP Economically Disconnected and Disinvested Areas}
\format{
A polygon \code{sf} object with 856 rows and 7
variables:
\describe{
\item{geoid_tract}{Unique 11-digit tract ID, assigned by the Census
Bureau. \strong{These correspond to tract boundaries from 2010, not 2020.}
Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the tract is in.
Character.}
\item{area_type}{Description of the tract's combined EDA and disinvested
status. Character.}
\item{eda}{Is the tract an Economically Disconnected Area (EDA)? Logical.}
\item{disinvested}{Is the tract a disinvested area? Logical.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
\href{https://datahub.cmap.illinois.gov/dataset/on-to-2050-layer-edas-disinvested-areas}{CMAP Data Hub (temporarily offline)}
}
\usage{
eda_sf
}
\description{
To advance \href{https://www.cmap.illinois.gov/2050}{ON TO 2050}'s commitment to
inclusive growth, CMAP has defined Economically Disconnected and Disinvested
Areas within the region. These classifications have been made at the Census
tract level, covering the 7-county CMAP region. Economically Disconnected
Areas (EDAs) are areas that describe population-based estimates of
concentrated low-income and minority status, whereas disinvested areas
outline places with long-term declines in employment and other markers of
commercial market weakness. Together, these communities experience a
persistent, long-term lack of market investment, leading to declining
property values, taxes, employment, and, frequently, population.
Disinvestment often constrains the ability of any individual community to
respond effectively to these losses, and high tax rates and low market
potential limit private investment.
}
\examples{
# Display the EDAs and disinvested areas with ggplot2
library(ggplot2)
ggplot(eda_sf) + geom_sf(aes(fill = area_type), lwd = 0.1) + theme_void()
}
\seealso{
\url{https://www.cmap.illinois.gov/2050/maps/eda}
}
\keyword{datasets}
