% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idot_sf}
\alias{idot_sf}
\title{IDOT Districts}
\format{
A polygon \code{sf} object with 9 rows and 4
variables:
\describe{
\item{district}{District ID, assigned by IDOT. Character.}
\item{region}{Region ID, assigned by IDOT. Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
counties, aggregated into
\href{https://idot.illinois.gov/about-idot/idot-regions/index}{IDOT Regions}
}
\usage{
idot_sf
}
\description{
The nine highway districts covering the entire state of Illinois, as defined
by the Illinois Department of Transportation (IDOT). Includes a column
indicating which of the five transportation regions each district belongs to.
Created using the county boundaries in the US Census Bureau's TIGER/Line
shapefiles, 2022 vintage.
}
\examples{
# Display the IDOT districts/regions with ggplot2
library(ggplot2)
ggplot(idot_sf) + geom_sf(aes(fill = region), lwd = 0.1) + theme_void()
}
\keyword{datasets}
