% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ilga_house_sf}
\alias{ilga_house_sf}
\title{Illinois State House Districts}
\format{
A polygon \code{sf} object with 118 rows and
5 variables:
\describe{
\item{dist_num}{Illinois State House District number. Integer.}
\item{dist_name}{Name of the district. Character.}
\item{cmap}{Does the district overlap the 7-county CMAP region? Logical.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
ilga_house_sf
}
\description{
The Illinois General Assembly House Districts. From the US Census Bureau's
TIGER/Line shapefiles, 2019 vintage.
}
\details{
Census Bureau description:

\emph{"State Legislative Districts (SLDs) are the areas from which members are
elected to state legislatures. The Census Bureau first reported data for SLDs
as part of the 2000 Public Law (P.L.) 94-171 Redistricting Data File.}

\emph{"Current SLDs (2010 Election Cycle) -- States participating in Phase 1 of
the 2010 Census Redistricting Data Program voluntarily provided the Census
Bureau with the 2006 election cycle boundaries, codes, and, in some cases,
names for their SLDs. All 50 states, plus the District of Columbia and Puerto
Rico, participated in Phase 1, State Legislative District Project (SLDP) of
the 2010 Census Redistricting Data Program. States subsequently provided
legal changes to those plans through the Redistricting Data Office and/or
corrections as part of Phase 2 of the 2010 Census Redistricting Data Program,
as needed.}

\emph{"The SLDs embody the upper (senate) and lower (house) chambers of the state
legislature. A unique three-character census code, identified by state
participants, is assigned to each SLD within a state. In Connecticut, Hawaii,
Illinois, Louisiana, Maine, Massachusetts, New Jersey, Ohio, and Puerto Rico,
state officials did not define the SLDs to cover all of the state or state
equivalent area (usually bodies of water). In these areas with no SLDs
defined, the code "ZZZ" has been assigned, which is treated within state as a
single SLD for purposes of data presentation."}

Note: The aforementioned "ZZZ" district, which comprises the Illinois portion
of Lake Michigan, has been excluded from this dataset.
}
\examples{
# Display the ILGA House Districts with ggplot2
library(ggplot2)
ggplot(data = ilga_house_sf) + geom_sf(aes(fill = cmap), lwd = 0.1) + theme_void()
}
\keyword{datasets}
