% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puma_sf}
\alias{puma_sf}
\title{Census Public Use Microdata Areas (PUMAs)}
\format{
A multipolygon \code{sf} object with 60 rows and 4
variables:
\describe{
\item{geoid_puma}{Unique 7-digit PUMA ID, assigned by the Census Bureau.
Character.}
\item{name}{A Census-assigned name to help describe a PUMA's extent.
Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
puma_sf
}
\description{
The Census PUMAs covering the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2023 vintage.
}
\details{
Census Bureau description:

\emph{"Public Use Microdata Areas (PUMAs) are geographic areas for which the
Census Bureau provides selected extracts of raw data from a small sample of
census records that are screened to protect confidentiality. These extracts
are referred to as public use microdata sample (PUMS) files.}

\emph{"PUMAs are statistical geographic areas defined for the tabulation and
dissemination of American Community Survey (ACS) and Puerto Rico Community
Survey Public Use Microdata Sample (PUMS) data as well as ACS period
estimates, and decennial census data. Nesting within states or equivalent
entities, PUMAs cover the entirety of the United States, Puerto Rico, Guam,
and the U.S. Virgin Islands. PUMA delineations are subject to population
threshold criteria, “building block” geography criteria, and geographic
nesting criteria. State Data Centers (SDCs) define PUMAs with the cooperation
of regional, state, local, and tribal governments, organizations, and other
interested data users. Each PUMA must have a minimum population of 100,000 at
the time of delineation and throughout the decade.}
}
\examples{
# Display the PUMAs with ggplot2
library(ggplot2)
ggplot(puma_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
