% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subzone_sf}
\alias{subzone_sf}
\title{CMAP Travel Modeling Subzones}
\format{
A polygon \code{sf} object with 17418 rows and 11
variables:
\describe{
\item{subzone17}{Numeric subzone ID. Integer.}
\item{zone17}{Numeric ID of parent zone, corresponding to \code{zone_sf}.
Integer.}
\item{capzone17}{The CMAP capacity zone characterization of the subzone.
Integer. \itemize{
\item{1 = Chicago Central Business District}
\item{2 = Remainder of Chicago Central Area}
\item{3 = Remainder of City of Chicago}
\item{4 = Inner ring suburbs where Chicago street grid is generally
maintained}
\item{5 = Remainder of IL portion of the Chicago Urbanized Area}
\item{6 = IN portion of the Chicago Urbanized Area}
\item{7 = Other Urbanized Areas and Urban Clusters within the CMAP
Metropolitan Planning Area plus other Urbanized Areas in IL}
\item{8 = Other Urbanized Areas and Urban Clusters in IN}
\item{9 = Remainder of CMAP Metropolitan Planning Area}
\item{10 = Remainder of Lake County, IN (rural)}
\item{11 = External area}
}}
\item{cmap}{Is subzone in the CMAP MPA? (The MPA includes the 7 CMAP
counties, plus Aux Sable Township in Grundy County and Sandwich & Somonauk
Townships in DeKalb County.) Logical.}
\item{chicago}{Is subzone in the City of Chicago? (Approximate, in some
cases.) Logical.}
\item{cbd}{Is subzone in the Chicago Central Business District? (CBD is
defined as the area bounded by Chicago Ave, Halsted St, Roosevelt Rd, and
Lake Michigan.) Logical.}
\item{township_range}{4-digit ID of the
\href{https://clearinghouse.isgs.illinois.edu/data/reference/illinois-plss-townships}{PLSS township}
the subzone is in; first 2 digits are township, last 2 are range. Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the subzone is
in; first 2 digits are the state FIPS code. Character.}
\item{state}{State the subzone is in (IL, IN or WI). Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
\href{https://datahub.cmap.illinois.gov/dataset/cmap-modeling-zone-systems}{CMAP Data Hub}
}
\usage{
subzone_sf
}
\description{
The 2017 subzones (a.k.a. MAZs) used by \href{http://cmap.illinois.gov}{CMAP} for
travel demand modeling and socioeconomic forecasting. Covers the 7-county
CMAP region, plus 14 other counties in Illinois, Indiana and Wisconsin (some
only partially). The subzones are derived from the Illinois Public Land
Survey System (PLSS), with most of the subzones within the CMAP region
corresponding to a PLSS quarter-section. This forms a fairly regular grid of
0.5 mile wide squares, although there are some irregularities.
}
\examples{
# Display the subzones with ggplot2
library(ggplot2)
ggplot(subzone_sf) + geom_sf(aes(fill = cmap), lwd = 0.1) + theme_void()
}
\keyword{datasets}
