% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{township_sf}
\alias{township_sf}
\title{Townships}
\format{
A multipolygon \code{sf} object with 128 rows and
5 variables:
\describe{
\item{geoid_cousub}{10-digit unique county subdivision/township ID,
assigned by the Census Bureau. Character.}
\item{township}{Township name. Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the township is
in. Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
township_sf
}
\description{
The political townships (also referred to as "county subdivisions" in Census
Bureau terminology) that are within the CMAP Metropolitan Planning Area
(MPA). (The MPA includes the 7 CMAP counties, plus Aux Sable Township in
Grundy County and Sandwich & Somonauk Townships in DeKalb County.) From the
US Census Bureau's TIGER/Line shapefiles, 2019 vintage.
}
\details{
Census Bureau description:

\emph{"County Subdivisions are the primary divisions of counties and equivalent
entities. They include census county divisions, census subareas, minor civil
divisions, and unorganized territories and can be classified as either legal
or statistical. Each county subdivision is assigned a five-character numeric
Federal Information Processing Series (FIPS) code based on alphabetical
sequence within state and an eight-digit National Standard feature
identifier."}

Note: The entire City of Chicago (other than the portion of O'Hare in DuPage
County) is included as a single township in this dataset, and has not been
subdivided into the eight theoretical townships defined by the Cook County
Clerk's Office for the purposes of collecting property tax.
}
\examples{
# Display the townships with ggplot2
library(ggplot2)
ggplot(data = township_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
