% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tract_sf}
\alias{tract_sf}
\title{Census Tracts}
\format{
A polygon \code{sf} object with 1983 rows and 4
variables:
\describe{
\item{geoid_tract}{Unique 11-digit tract ID, assigned by the Census
Bureau. Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the tract is in.
Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
tract_sf
}
\description{
The Census Tracts within the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2019 vintage.
}
\details{
Census Bureau description:

\emph{"Census Tracts are small, relatively permanent statistical subdivisions of a
county or equivalent entity that are updated by local participants prior to
each decennial census as part of the Census Bureau's Participant Statistical
Areas Program. The Census Bureau delineates census tracts in situations where
no local participant existed or where state, local, or tribal governments
declined to participate. The primary purpose of census tracts is to provide a
stable set of geographic units for the presentation of statistical data.}

\emph{"Census tracts generally have a population size between 1,200 and 8,000
people, with an optimum size of 4,000 people. A census tract usually covers a
contiguous area; however, the spatial size of census tracts varies widely
depending on the density of settlement. Census tract boundaries are
delineated with the intention of being maintained over a long time so that
statistical comparisons can be made from census to census. Census tracts
occasionally are split due to population growth or merged as a result of
substantial population decline.}

\emph{"Census tract boundaries generally follow visible and identifiable features.
They may follow nonvisible legal boundaries, such as minor civil division
(MCD) or incorporated place boundaries in some states and situations, to
allow for census-tract-to-governmental-unit relationships where the
governmental boundaries tend to remain unchanged between censuses. State and
county boundaries always are census tract boundaries in the standard census
geographic hierarchy. Tribal census tracts are a unique geographic entity
defined within federally recognized American Indian reservations and
off-reservation trust lands and can cross state and county boundaries. Tribal
census tracts may be completely different from the census tracts and block
groups defined by state and county."}
}
\examples{
# Display the tracts with ggplot2
library(ggplot2)
ggplot(tract_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
