% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward_sf}
\alias{ward_sf}
\title{Chicago Wards}
\format{
A multipolygon \code{sf} object with 50 rows and 3
variables:
\describe{
\item{ward_num}{Ward number. Integer.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
\href{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Wards-2015-/sp34-6z76}{Chicago Data Portal}
}
\usage{
ward_sf
}
\description{
The official boundaries of the Chicago wards established in May of 2015.
Obtained 2/12/2023. Updated boundaries will be provided once the 2023 wards
come into effect in May 2023.
}
\examples{
# Display the wards with ggplot2
library(ggplot2)
ggplot(ward_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
