% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_tract2zone}
\alias{xwalk_tract2zone}
\title{Tract-to-Zone Crosswalk}
\format{
A tibble with 6910 rows and 5
variables:
\describe{
\item{geoid_tract}{Unique 11-digit tract ID, assigned by the Census Bureau.
Corresponds to \code{tract_sf} (although that only includes the tracts in the
7-county CMAP region). Character.}
\item{zone17}{Numeric zone ID. Corresponds to \code{zone_sf}. Integer.}
\item{hu_pct}{Proportion of the tract's housing units (occupied or vacant)
located in the specified zone. Multiply this by a tract-level measure of a
housing attribute (e.g. vacant homes) to estimate the zone's portion.
Double.}
\item{hh_pct}{Proportion of the tract's households (i.e. occupied housing
units) living in the specified zone. Multiply this by a tract-level measure
of a household attribute (e.g. car-free households) to estimate the zone's
portion. Double.}
\item{pop_pct}{Proportion of the tract's total population (including group
quarters) living in the specified zone. Multiply this by a tract-level
measure of a population attribute (e.g. race/ethnicity) to estimate the
zone's portion. Double.}
}
}
\usage{
xwalk_tract2zone
}
\description{
This table contains a set of factors to apportion Census tract-level data
among the CMAP travel modeling zones. Separate factors are provided for
apportioning housing unit, household, and population attributes. All factors
were determined by calculating the percentage of a tract's housing units,
households and population that were located in each of its component blocks,
according to the 2010 Decennial Census, and then assigning each block to a
zone (based on the location of the block's centroid point). Zones that
do not contain the centroid of any blocks with at least one housing unit,
household or person are not present in this table, and should be considered
unpopulated.
}
\details{
Other than in certain areas of Chicago, tracts tend to be larger than zones
and have highly irregular boundaries, so in most cases the population,
households and/or housing units in a tract are split across multiple zones.
For that reason, it is not appropriate to use a one-to-one tract-to-zone
assignment to apportion Census data among zones, and this crosswalk should be
used instead.

To use this crosswalk effectively, Census data should be joined to it (not
vice versa, since tract IDs appear multiple times in this table). Once the
data is joined, it should be multiplied by the appropriate factor (depending
whether the data of interest is measured at the housing unit, household or
person level), and then the result should be summed by zone ID. If
calculating rates, this should only be done after the counts have been summed
to zone. The resulting table can then be joined to \code{zone_sf} for mapping, if
desired.

If your data is also available at the block group level, it is recommended
that you use that with \code{xwalk_blockgroup2zone} instead of the tract-level
allocation. If the zone geography is too coarse for your needs, you can use
subzones instead with \code{xwalk_tract2subzone} or \code{xwalk_blockgroup2subzone}.
}
\examples{
# View the tract allocations for zone17 == 55
dplyr::filter(xwalk_tract2zone, zone17 == 55)

# Map the zones missing from xwalk_tract2zone (i.e. no HU/HH/pop)
library(ggplot2)
ggplot(dplyr::anti_join(zone_sf, xwalk_tract2zone)) +
  geom_sf(fill = "red", lwd = 0.1) +
  geom_sf(data = zone_sf, fill = NA, lwd = 0.1) +
  theme_void()
}
\keyword{datasets}
