% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_sf}
\alias{zcta_sf}
\title{Census ZIP Code Tabulation Areas (ZCTAs)}
\format{
A multipolygon \code{sf} object with 320 rows and 3
variables:
\describe{
\item{geoid_zcta}{Unique 5-digit ZCTA ID, corresponding to a 5-digit USPS
ZIP Code. Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
zcta_sf
}
\description{
The Census ZCTAs covering the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2023 vintage.
}
\details{
Census Bureau description:

\emph{"ZIP Code Tabulation Areas (ZCTAs) are approximate area representations of
U.S. Postal Service (USPS) five-digit ZIP Code service areas that the Census
Bureau creates using whole blocks to present statistical data from censuses
and surveys. The Census Bureau defines ZCTAs by allocating each block that
contains addresses to a single ZCTA, usually to the ZCTA that reflects the
most frequently occurring ZIP Code for the addresses within that tabulation
block. Blocks that do not contain addresses but are completely surrounded by
a single ZCTA (enclaves) are assigned to the surrounding ZCTA; those
surrounded by multiple ZCTAs will be added to a single ZCTA based on limited
buffering performed between multiple ZCTAs. The Census Bureau identifies
five-digit ZCTAs using a five-character numeric code that represents the most
frequently occurring USPS ZIP Code within that ZCTA, and this code may
contain leading zeros. Not all ZIP Codes in use by the USPS may have a ZCTA
delineated to represent them, The USPS makes periodic changes to ZIP Codes to
support more efficient mail delivery. In addition, the ZCTA delineation
process primarily uses residential addresses and has a bias towards ZIP Codes
used for city-style mail delivery, thus there may be ZIP Codes that are
primarily nonresidential or used for PO boxes only that may not have a
corresponding ZCTA. ZIP Code is a trademark of the U.S. Postal Service."}
}
\examples{
# Display the ZCTAs with ggplot2
library(ggplot2)
ggplot(zcta_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
