% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zone_sf}
\alias{zone_sf}
\title{CMAP Travel Modeling Zones, 2017}
\format{
A polygon \code{sf} object with 3632 rows and 9
variables:
\describe{
\item{zone17}{Numeric zone ID. Integer.}
\item{cmap}{Is zone in the CMAP MPO area? (The MPO area includes the 7
CMAP counties, plus Aux Sable Township in Grundy County and Sandwich &
Somonauk Townships in DeKalb County.) Logical.}
\item{chicago}{Is zone in the City of Chicago? (Approximate, in some
cases.) Logical.}
\item{cbd}{Is zone in the Chicago Central Business District? (CBD is
defined as the area bounded by Chicago Ave, Halsted St, Roosevelt Rd, and
Lake Michigan.) Logical.}
\item{township_range}{4-digit ID of the
\href{https://clearinghouse.isgs.illinois.edu/data/reference/illinois-plss-townships}{PLSS township}
the zone is in; first 2 digits are township, last 2 are range. Character.}
\item{county_fips}{5-digit FIPS code of the county the zone is in; first
2 digits are the state FIPS code. Character.}
\item{state}{State the zone is in (IL, IN or WI). Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} polygon.}
}
}
\source{
\href{https://datahub.cmap.illinois.gov/dataset/cmap-modeling-zone-systems/resource/8dd37215-98dc-4244-9623-2d28c4e1611c}{CMAP Data Hub}
}
\usage{
zone_sf
}
\description{
The 2017 zones (a.k.a. TAZs) used by \href{http://cmap.illinois.gov}{CMAP} for
travel demand modeling. Covers the 7-county CMAP region, plus 14 other
counties in Illinois, Indiana and Wisconsin (some only partially). Zones are
aggregations of subzones (\code{subzone_sf}).
}
\examples{
# Display the zones with ggplot2
library(ggplot2)
ggplot(data = zone_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
