% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_sf}
\alias{block_sf}
\title{Census Blocks}
\format{
\code{block_sf} is a multipolygon \code{sf} object with 144082 rows
and 4 variables:
\describe{
\item{geoid_block}{Unique 15-digit block ID, assigned by the Census Bureau.
The parent tract and block group can be identified from the first 11 and 12
digits, respectively. Character.}
\item{county_fips}{Unique 5-digit FIPS code of the county the block is in.
Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
block_sf
}
\description{
The Census Blocks within the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2022 vintage. \strong{Use \code{block_sf} for data from the 2020 decennial census or the
American Community Survey (ACS) from 2020 onward.}
}
\details{
Census Bureau description:

\emph{"Blocks (Census Blocks or Tabulation Blocks) are statistical areas bounded
by visible features, such as streets, roads, streams, and railroad tracks,
and by nonvisible boundaries, such as selected property lines and city,
township, school district, and county limits and short line-of-sight
extensions of streets and roads. Generally, blocks are small in area; for
example, a city block bounded on all sides by streets. Blocks in suburban and
rural areas may be larger, more irregular in shape, and bounded by a variety
of features, such as roads, streams, and transmission lines. In remote areas,
blocks may even encompass hundreds of square miles. Blocks cover the entire
territory of the United States, Puerto Rico, and the Island Areas. Blocks
nest within all other tabulated census geographic entities at the time of the
decennial census and are the basis for all tabulated data from that census.
Census Block Numbers—Blocks are numbered uniquely with a four-digit census
block number from 0000 to 9999 within census tract, which nest within state
and county. The first digit of the census block number identifies the block
group. Block numbers beginning with a zero (in Block Group 0) are intended to
include only water area, but not all water-only blocks have block numbers
beginning with 0 (zero)."}
}
\examples{
# Display the blocks with ggplot2
library(ggplot2)
ggplot(block_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
