% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cca_sf}
\alias{cca_sf}
\title{Chicago Community Areas (CCAs)}
\format{
A multipolygon \code{sf} object with 77 rows and 4
variables:
\describe{
\item{cca_name}{Community area name. Character.}
\item{cca_num}{Numeric ID, as assigned by the City of Chicago. Integer.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
\href{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6}{Chicago Data Portal}
}
\usage{
cca_sf
}
\description{
The official boundaries of the Chicago Community Areas. Obtained 3/24/2021.
}
\examples{
# Display the CCAs with ggplot2
library(ggplot2)
ggplot(cca_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
