% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puma_sf}
\alias{puma_sf}
\title{Census Public Use Microdata Areas (PUMAs)}
\format{
A multipolygon \code{sf} object with 58 rows and 4
variables:
\describe{
\item{geoid_puma}{7-digit unique PUMA ID, assigned by the Census Bureau.
Character.}
\item{name}{A Census-assigned name to help describe a PUMA's extent.
Character.}
\item{sqmi}{Area in square miles. Double.}
\item{geometry}{Feature geometry. \code{sf} multipolygon.}
}
}
\source{
US Census Bureau
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
}
\usage{
puma_sf
}
\description{
The Census PUMAs covering the 7-county Chicago Metropolitan Agency for
Planning (CMAP) region. From the US Census Bureau's TIGER/Line shapefiles,
2019 vintage.
}
\details{
Census Bureau description: "Public Use Microdata Areas (PUMAs) are geographic
areas for which the Census Bureau provides selected extracts of raw data from
a small sample of census records that are screened to protect
confidentiality. These extracts are referred to as public use microdata
sample (PUMS) files.

"For the 2010 Census, each state, the District of Columbia, Puerto Rico, and
some Island Area participants delineated PUMAs for use in presenting PUMS
data based on a 5 percent sample of decennial census or American Community
Survey data. These areas are required to contain at least 100,000 people.
This is different from Census 2000 when two types of PUMAs were defined: a 5
percent PUMA as for 2010 and an additional super-PUMA designed to provide a 1
percent sample. The PUMAs are identified by a five-digit census code unique
within state."
}
\examples{
# Display the PUMAs with ggplot2
library(ggplot2)
ggplot(data = puma_sf) + geom_sf(lwd = 0.1) + theme_void()
}
\keyword{datasets}
