% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward_sf}
\alias{ward_sf}
\title{Chicago Wards}
\format{
A multipolygon \code{sf} object with 50 rows and
2 variables:
\describe{
\item{ward_num}{Ward number. Integer.}
}
}
\source{
\href{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Wards-2015-/sp34-6z76}{Chicago Data Portal}
}
\usage{
ward_sf
}
\description{
The official boundaries of the current Chicago Wards (established in May of
2015). Obtained 3/24/2021.
}
\examples{
# Display the wards with ggplot2
library(ggplot2)
ggplot(data = ward_sf) + geom_sf() + theme_void()
}
\keyword{datasets}
