% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_blockgroup2cca}
\alias{xwalk_blockgroup2cca}
\title{Block Group-to-CCA Crosswalk}
\format{
A tibble with 2180 rows and
5 variables:
\describe{
\item{geoid_blkgrp}{Unique 12-digit block group ID, assigned by the Census
Bureau. Corresponds to \code{blockgroup_sf}. Character.}
\item{cca_num}{Numeric CCA ID, as assigned by the City of Chicago.
Corresponds to \code{cca_sf}. Integer.}
\item{hu_pct}{Proportion of the block group's housing units (occupied or
vacant) located in the specified CCA. Multiply this by a block group-level
measure of a housing attribute (e.g. vacant homes) to estimate the CCA's
portion. Double.}
\item{hh_pct}{Proportion of the block group's households (i.e. occupied
housing units) living in the specified CCA. Multiply this by a block
group-level measure of a household attribute (e.g. car-free households) to
estimate the CCA's portion.Double.}
\item{pop_pct}{Proportion of the block group's total population (including
group quarters) living in the specified CCA. Multiply this by a block
group-level measure of a population attribute (e.g. race/ethnicity) to
estimate the CCA's portion. Double.}
}
}
\usage{
xwalk_blockgroup2cca
}
\description{
This table contains a set of factors to apportion Census block group-level
data among Chicago Community Areas (CCAs). Seperate factors are provided for
apportioning housing unit, household, and population attributes. All factors
were determined by calculating the percentage of a block group's housing
units, households and population that were located in each of its component
blocks, according to the 2010 Decennial Census, and then assigning each block
to a CCA (based on the location of the block's centroid point).
}
\details{
Generally speaking, block group boundaries align neatly with CCA boundaries
as they tend to follow similar features (e.g. rivers, major roads, rail
lines) but there are cases where the population, households and/or housing
units in a block group are split across multiple CCAs, or else are partially
within the City of Chicago and partially outside of it. For that reason, it
is not appropriate to use a one-to-one block group-to-CCA assignment to
apportion Census data among CCAs, and this crosswalk (or \code{xwalk_tract2cca}
for tract-level data) should be used instead.

To use this crosswalk effectively, Census data should be joined to it (not
vice versa, since block group IDs appear multiple times in this table). Once
the data is joined, it should be multiplied by the appropriate factor
(depending whether the data of interest is measured at the housing unit,
household or person level), and then the result should be summed by CCA. If
calculating rates, this should only be done after the counts have been summed
to CCA. The resulting table can then be joined to \code{cca_sf} for mapping, if
desired.

If your data is only available at the tract level, you can use
\code{xwalk_tract2cca} for a tract-level allocation instead.
}
\examples{
suppressPackageStartupMessages(library(dplyr))

# View the block groups with households not fully contained in a single CCA
filter(xwalk_blockgroup2cca, hh_pct < 1)

# Estimate CCA-level unemployment rate from block group-level ACS data
df_blkgrp <- tidycensus::get_acs(
  "block group", state = "IL", county = "031", table = "B23025",
  year = 2019, survey = "acs5", output = "wide", cache_table = TRUE
) \%>\%
  rename(civ_lf = B23025_003E, unemp = B23025_005E) \%>\%
  select(GEOID, civ_lf, unemp)

df_cca <- xwalk_blockgroup2cca \%>\%
  left_join(df_blkgrp, by = c("geoid_blkgrp" = "GEOID")) \%>\%
  mutate(civ_lf = civ_lf * pop_pct,
         unemp = unemp * pop_pct) \%>\%
  group_by(cca_num) \%>\%
  summarize_at(vars(civ_lf, unemp), sum) \%>\%
  mutate(unemp_rate = unemp / civ_lf)
df_cca

# Join to cca_sf for mapping
library(ggplot2)
cca_sf \%>\%
  left_join(df_cca, by = "cca_num") \%>\%
  ggplot() +
    geom_sf(aes(fill = unemp_rate), lwd = 0.1) +
    scale_fill_viridis_c(direction = -1) +
    theme_void()
}
\keyword{datasets}
