% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_blockgroup2zone}
\alias{xwalk_blockgroup2zone}
\alias{xwalk_blockgroup2zone_2010}
\title{Block Group-to-Zone Crosswalk}
\format{
\code{xwalk_blockgroup2zone} is a tibble with 13289 rows
and 6 variables:
\describe{
\item{geoid_blkgrp}{Unique 12-digit block group ID, assigned by the Census
Bureau. Corresponds to \code{blockgroup_sf} (although that only includes the
block groups in the 7-county CMAP region). Character.}
\item{zone17}{Numeric zone ID. Corresponds to \code{zone_sf}. Integer.}
\item{hu_pct}{Proportion of the block group's housing units (occupied or
vacant) located in the specified zone. Multiply this by a block group-level
measure of a housing attribute (e.g. vacant homes) to estimate the zone's
portion. Double.}
\item{hh_pct}{Proportion of the block group's households (i.e. occupied
housing units) living in the specified zone. Multiply this by a block
group-level measure of a household attribute (e.g. car-free households) to
estimate the zone's portion. Double.}
\item{pop_pct}{Proportion of the block group's total population (including
group quarters) living in the specified zone. Multiply this by a block
group-level measure of a population attribute (e.g. race/ethnicity) to
estimate the zone's portion. Double.}
\item{emp_pct}{Proportion of the block group's total jobs located in the
specified zone. Multiply this by a block group-level measure of an
employment attribute (e.g. retail jobs) to estimate the zone's portion.
\strong{Not available in \code{xwalk_blockgroup2zone_2010}.} Double.}
}

\code{xwalk_blockgroup2zone_2010} is a tibble with
12510 rows and
5 variables (no \code{emp_pct}).
}
\usage{
xwalk_blockgroup2zone

xwalk_blockgroup2zone_2010
}
\description{
This table contains a set of factors to apportion Census block group-level
data among the CMAP travel modeling zones. Separate factors are provided for
apportioning housing unit, household, population and employment attributes.
All factors were determined by calculating the percentage of a block group's
housing units, households, population and employment that were located in
each of its component blocks, according to the 2020 Decennial Census and 2019
LEHD, and then assigning each block to a zone (based on the location of the
block's centroid point). Zones that do not contain the centroid of any blocks
with at least one housing unit, household, person or job are \emph{not} present in
this table. \strong{Use \code{xwalk_blockgroup2zone} for data from the 2020 decennial
census or the American Community Survey (ACS) from 2020 onward. For data from
the 2010 decennial census or ACS from 2010 through 2019, use
\code{xwalk_blockgroup2zone_2010}.}
}
\details{
Other than in certain areas of Chicago, block groups tend to be larger than
zones and have highly irregular boundaries, so in most cases the jobs,
population, households and/or housing units in a block group are split across
multiple zones. For that reason, it is not appropriate to use a one-to-one
block group-to-zone assignment to apportion Census data among zones, and this
crosswalk should be used instead.

To use this crosswalk effectively, Census data should be joined to it (not
vice versa, since block group IDs appear multiple times in this table). Once
the data is joined, it should be multiplied by the appropriate factor
(depending whether the data of interest is measured at the housing unit,
household, person or job level), and then the result should be summed by zone
ID. If calculating rates, this should only be done after the counts have been
summed to zone. The resulting table can then be joined to \code{zone_sf} for
mapping, if desired.

If your data is only available at the tract level, you can use
\code{xwalk_tract2zone} for a tract-level allocation instead. If the zone
geography is too coarse for your needs, you can use subzones instead with
\code{xwalk_blockgroup2subzone} or \code{xwalk_tract2subzone}.
}
\examples{
# View the block group allocations for zone17 == 55
dplyr::filter(xwalk_blockgroup2zone, zone17 == 55)

# Map the zones missing from xwalk_blockgroup2zone (i.e. no HU/HH/pop/emp)
library(ggplot2)
ggplot(dplyr::anti_join(zone_sf, xwalk_blockgroup2zone)) +
  geom_sf(fill = "red", lwd = 0.1) +
  geom_sf(data = zone_sf, fill = NA, lwd = 0.1) +
  theme_void()
}
\keyword{datasets}
